Moved sources of FailableMemoryAllocator to TestMemoryAllocator
