+++ CppUTest.vcproj
@@ -227,28 +227,6 @@
 				</FileConfiguration>
 			</File>
 			<File
-				RelativePath="SRC\CPPUTEST\FailableMemoryAllocator.cpp"
-				>
-				<FileConfiguration
-					Name="Debug|Win32"
-					>
-					<Tool
-						Name="VCCLCompilerTool"
-						AdditionalIncludeDirectories=""
-						PreprocessorDefinitions=""
-					/>
-				</FileConfiguration>
-				<FileConfiguration
-					Name="Release|Win32"
-					>
-					<Tool
-						Name="VCCLCompilerTool"
-						AdditionalIncludeDirectories=""
-						PreprocessorDefinitions=""
-					/>
-				</FileConfiguration>
-			</File>
-			<File
 				RelativePath="SRC\CPPUTEST\JUnitTestOutput.cpp"
 				>
 				<FileConfiguration
