+++ Makefile.am
@@ -49,7 +49,6 @@ lib_libCppUTest_a_SOURCES = \
 	src/CppUTest/TestRegistry.cpp \
 	src/CppUTest/TestResult.cpp \
 	src/CppUTest/Utest.cpp \
-	src/CppUTest/FailableMemoryAllocator.cpp \
 	src/Platforms/$(CPP_PLATFORM)/UtestPlatform.cpp
 
 include_cpputestdir = $(includedir)/CppUTest
@@ -79,8 +78,7 @@ include_cpputest_HEADERS = \
 	include/CppUTest/TestResult.h \
 	include/CppUTest/TestTestingFixture.h \
 	include/CppUTest/Utest.h \
-	include/CppUTest/UtestMacros.h \
-	include/CppUTest/FailableMemoryAllocator.h
+	include/CppUTest/UtestMacros.h
 
 lib_libCppUTestExt_a_CPPFLAGS = $(lib_libCppUTest_a_CPPFLAGS)
 lib_libCppUTestExt_a_CFLAGS = $(lib_libCppUTest_a_CFLAGS)
