--- src/CppUTest/FailableMemoryAllocator.cpp
@@ -1,73 +1,0 @@
-/*
- * Copyright (c) 2007, Michael Feathers, James Grenning and Bas Vodde
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in the
- *       documentation and/or other materials provided with the distribution.
- *     * Neither the name of the <organization> nor the
- *       names of its contributors may be used to endorse or promote products
- *       derived from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE EARLIER MENTIONED AUTHORS ``AS IS'' AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL <copyright holder> BE LIABLE FOR ANY
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-#include "CppUTest/TestHarness.h"
-#include "CppUTest/FailableMemoryAllocator.h"
-#include "CppUTest/PlatformSpecificFunctions.h"
-
-
-FailableMemoryAllocator::FailableMemoryAllocator(const char* name_str, const char* alloc_name_str, const char* free_name_str)
-: TestMemoryAllocator(name_str, alloc_name_str, free_name_str)
-, toFailCount_(0), currentAllocNumber_(0)
-{
-    PlatformSpecificMemset(allocsToFail_, 0, sizeof(allocsToFail_));
-}
-
-void FailableMemoryAllocator::failAllocNumber(int number)
-{
-    if (toFailCount_ >= MAX_NUMBER_OF_FAILED_ALLOCS)
-        FAIL("Maximum number of failed memory allocations exceeded");
-    allocsToFail_[toFailCount_++] = number;
-}
-
-char* FailableMemoryAllocator::alloc_memory(size_t size, const char* file, int line)
-{
-    currentAllocNumber_++;
-    if (shouldBeFailedAlloc_())
-        return NULL;
-    return TestMemoryAllocator::alloc_memory(size, file, line);
-}
-
-bool FailableMemoryAllocator::shouldBeFailedAlloc_()
-{
-    for (int i = 0; i < toFailCount_; i++)
-        if (currentAllocNumber_ == allocsToFail_[i])
-            return true;
-    return false;
-}
-
-char* FailableMemoryAllocator::allocMemoryLeakNode(size_t size)
-{
-    return (char*)PlatformSpecificMalloc(size);
-}
-
-void FailableMemoryAllocator::clearFailedAllocations()
-{
-    toFailCount_ = 0;
-    currentAllocNumber_ = 0;
-    PlatformSpecificMemset(allocsToFail_, 0, sizeof(allocsToFail_));
-}
-
